更新

+++ app/views/user.py
@@ -59,7 +59,7 @@ class UserReource(Resource):
         return '保存成功'
 
     def put(self):
-        self.rp.add_argument('id', type=str, location='json')
+        self.rp.add_argument('id', type=int, location='json', required=True)
         self.rp.add_argument('name', type=str, location='json')
         self.rp.add_argument('age', type=int, location='json')
         self.rp.add_argument('gender', type=str, location='json')
@@ -67,6 +67,7 @@ class UserReource(Resource):
         self.rp.add_argument('mobile', location='json', default=None)
 
         new_rp = self.rp.parse_args()
+
         try:
             User.query.filter(User.id == new_rp.id).update({'name': new_rp.name,
                                                             'age': new_rp.age,
@@ -75,73 +76,15 @@ class UserReource(Resource):
                                                             'mobile': new_rp.mobile
                                                             })
             User.commit()
+            return '修改成功'
         except:
             return '修改失败'
-        return '修改成功'
 
     def delete(self):
-        self.rp.add_argument('id', type=str, location='json')
+        self.rp.add_argument('id', type=int, location='json', required=True)
         new_rp = self.rp.parse_args()
-        User.query.filter(User.id == new_rp.id).delete()
-        User.commit()
+        if new_rp.id:
+            User.query.filter(User.id == new_rp.id).delete()
+            User.commit()
         return '删除成功'
 
-
-class UserReource1(Resource):
-    """用户表,增删改查"""
-
-    def __init__(self):
-        self.rp = RequestParser()
-
-    def get(self):
-        self.rp.add_argument('id', required=True, type=int, location='args')
-        new_rp = self.rp.parse_args()
-
-        user = User.query.filter(User.id == new_rp.id).first()
-        if user:
-            return {"username": user.name, "age": user.age}
-        else:
-            return {"username": None}
-
-    def post(self):
-        self.rp.add_argument('name', type=str, location='json')
-        self.rp.add_argument('age', type=int, location='json')
-        self.rp.add_argument('gender', type=str, location='json')
-        self.rp.add_argument('addr', default=None, location='json')
-        self.rp.add_argument('mobile', location='json', default=None)
-
-        new_rp = self.rp.parse_args()
-
-        User(name=new_rp.name,
-             age=new_rp.age,
-             gender=new_rp.gender,
-             addr=new_rp.addr,
-             mobile=new_rp.mobile,
-             create_time=datetime.now().strftime("%Y-%m-%d %H:%M:%S")
-             ).save()
-        return '保存成功'
-
-    def put(self):
-        self.rp.add_argument('id', type=str, location='json')
-        self.rp.add_argument('name', type=str, location='json')
-        self.rp.add_argument('age', type=int, location='json')
-        self.rp.add_argument('gender', type=str, location='json')
-        self.rp.add_argument('addr', default=None, location='json')
-        self.rp.add_argument('mobile', location='json', default=None)
-
-        new_rp = self.rp.parse_args()
-        User.query.filter(User.id == new_rp.id).update({'name': new_rp.name,
-                                                        'age': new_rp.age,
-                                                        'gender': new_rp.gender,
-                                                        'addr': new_rp.addr,
-                                                        'mobile': new_rp.mobile
-                                                        })
-        User.commit()
-        return '修改成功'
-
-    def delete(self):
-        self.rp.add_argument('id', type=str, location='json')
-        new_rp = self.rp.parse_args()
-        User.query.filter(User.id == new_rp.id).delete()
-        User.commit()
-        return '删除成功'
